Fix
 * git2go/blob.go:63: cannot use _Ctype_size_t(len(goBuf)) (type C.size_t) as type C.ulong in function argument
 * git2go/merge.go:169 git_merge_analysis: now have five arguments

+++ blob.go
@@ -61,7 +61,7 @@ func blobChunkCb(buffer *C.char, maxLen C.size_t, payload unsafe.Pointer) int {
 		data.Error = err
 		return -1
 	}
-	C.memcpy(unsafe.Pointer(buffer), unsafe.Pointer(&goBuf[0]), C.size_t(len(goBuf)))
+	C.memcpy(unsafe.Pointer(buffer), unsafe.Pointer(&goBuf[0]), C.ulong(len(goBuf)))
 	return len(goBuf)
 }
